fix(writer): address spec review — sponsor validation, epoch increment, type annotations
- Sponsor must be an ACTIVE/PROBATION writer (Spec C2 §4.4)
- Router increments EpochTracker after successful gated dispatch
- WriterRecord.enrolled_at and approved_at typed as int (not float)
- Updated all tests to use registered ACTIVE sponsors
- Added test_inactive_sponsor and test_unregistered_sponsor assertions

+++ src/ltp/execution/writer.py
@@ -236,10 +236,10 @@ class WriterRecord:
     identity:           WriterIdentity
     state:              WriterState
     approval_path:      ApprovalPath
-    enrolled_at:        float                   # Unix timestamp of enrollment
+    enrolled_at:        int                     # Timestamp in milliseconds
 
     # Optional fields set on approval / progression
-    approved_at:        Optional[float]         = None
+    approved_at:        Optional[int]           = None
     approved_by:        Optional[bytes]         = None  # 32-byte actor fingerprint
     sponsors:           list[bytes]             = field(default_factory=list)
